Separate test cases into unit/functional
