+++ Jenkinsfile
@@ -47,6 +47,7 @@ pipeline {
 					dir('./framework') { sh "npm run test:unit" }
 					dir('./services/blockchain-connector') { sh "npm run test:unit" }
 					dir('./services/blockchain-indexer') { sh "npm run test:unit" }
+					dir('./services/fee-estimator') { sh "npm run test:unit" }
 					dir('./services/core') { sh "npm run test:unit" }
 					dir('./services/market') { sh "npm run test:unit" }
 					dir('./services/newsfeed') { sh "npm run test:unit" }
